+++ src/interface_type_001_actuators.h
@@ -21,7 +21,7 @@ struct actuator_s
    char arduino_pin;
    char arduino_pin_type;
    
-   int old_val;
+   int16_t old_val;
 };
 
 void _interface_type_001_free_actuators_queue_elem(void *d);
