+++ src/interface_type_001_sensors.c
@@ -226,7 +226,7 @@ int16_t interface_type_001_sensors_process_traps(int16_t numTrap, char *data, in
 
 
 
-struct sensor_s *interface_type_001_sensors_valid_and_malloc_sensor(int id_sensor_actuator, char *name, char *parameters)
+struct sensor_s *interface_type_001_sensors_valid_and_malloc_sensor(int16_t id_sensor_actuator, char *name, char *parameters)
 {
    int pin_id;
    int type_id;
