+++ src/interface_type_001_sensors.h
@@ -15,7 +15,7 @@
 
 struct sensor_s
 {
-   int sensor_id;
+   uint16_t sensor_id;
    char name[20];
    
    char arduino_pin_type;
@@ -34,18 +34,14 @@ struct sensor_s
    
    mea_timer_t timer;
    
-   long *nbtrap;
+   uint32_t *nbtrap;
 };
 
 void interface_type_001_sensors_free_queue_elem(void *d);
-//mea_error_t interface_type_001_sensors_process_traps(int numTrap, void *args, char *buff);
 int16_t interface_type_001_sensors_process_traps(int16_t numTrap, char *data, int16_t l_data, void *userdata);
 
-struct sensor_s *interface_type_001_sensors_valid_and_malloc_sensor(int id_sensor_actuator, char *name, char *parameters);
-//mea_error_t interface_type_001_sensors_process_xpl_msg(interface_type_001_t *i001, xPL_ServicePtr theService, xPL_NameValueListPtr ListNomsValeursPtr, char *device, char *type);
+struct sensor_s *interface_type_001_sensors_valid_and_malloc_sensor(int16_t id_sensor_actuator, char *name, char *parameters);
 mea_error_t interface_type_001_sensors_process_xpl_msg(interface_type_001_t *i001, xPL_ServicePtr theService, xPL_MessagePtr msg, char *device, char *type);
-
-//int16_t interface_type_001_sensors_poll_inputs(interface_type_001_t *i001, tomysqldb_md_t *md);
 int16_t interface_type_001_sensors_poll_inputs(interface_type_001_t *i001);
 void interface_type_001_sensors_init(interface_type_001_t *i001);
